[ACE 6-TAO 2] Fixed warnings

+++ ACE/ace/ARGV.cpp
@@ -351,7 +351,7 @@ ACE_ARGV_T<CHAR_TYPE>::create_buf_from_queue (void)
                         -1);
 #else
   ACE_NEW_RETURN (this->buf_,
-                  CHAR_TYPE[this->length_ + this->argc_],
+                  CHAR_TYPE[this->length_ + static_cast<unsigned int> (this->argc_)],
                   -1);
 #endif /* ACE_HAS_ALLOC_HOOKS */
